backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
motor
pymongo
pydantic
pydantic-settings
email-validator
python-jose
passlib[bcrypt]
httpx
pyjhora
python-dateutil
pyswisseph
geocoder
timezonefinder
geopy